requirements.txt
Flask==3.0.3
flask-pymongo==2.3.0
python-dotenv==1.0.1
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
numpy==1.26.4
Pillow==10.1.0
werkzeug
pandas==2.2.2
